Tests has now their own files in js/c folders
